AuthGuard relocated inside a folder
